refactor(farmer): split FarmerLOCC (688→147) em componentes por aba
Extrai os sub-componentes internos (já memoizados) para src/components/farmer/locc/:
OverviewTab, ExperimentsTab, CapacityTab, AdaptiveTab (cada aba dona dos próprios
hooks), ExperimentCard, NewExperimentDialog, primitivos (TabSkeleton/MetricRow/
WeightBar), helpers (fmt/fmtDur/cores) e types. O pai vira shell + wiring de tabs
com lazy-mount por visita. Comportamento preservado 1:1; +23 testes.
